In background.html, added switching tooltip text to "Reddited" or "not Reddited" depending on whether or not the current URL is submitted. (This appears regardless of the setting of the status badges option)
Also in background.html, changed "?" status badge to "N"

In options.html, changed the setting description from "?" to "N"

Updated popup.js so that when the url has been submitted today, it says simply "today" instead of "1 days ago"

In manifest.json, rewrote the description and removed the default badgeText and bumped the version number.

In README, added the new description.